perf: extract dashboard-data.js to avoid ink import in non-TTY paths
Move pure data functions (formatAge, computeSummary, getDashboardData,
renderPlainDashboard) into lib/ui/dashboard-data.js which has no ink
dependency. This eliminates the slow es-toolkit/compat transitive import
(~30s per process) for:

- bin/rally.js --json and non-TTY dashboard modes
- test/non-tty.test.js (moved out of test/ui/ since it doesn't need ink)
- test/integration.test.js dashboard tests

UI tests that need ink-testing-library still pay the import cost but
test correctness is unaffected.
