+++ lib/ui/Dashboard.jsx
@@ -1,55 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { Box, Text, useInput, useApp } from 'ink';
-import { existsSync } from 'node:fs';
-import { getActiveDispatches } from '../active.js';
 import DispatchTable from './components/DispatchTable.jsx';
-import { formatAge } from './components/DispatchTable.jsx';
+import { computeSummary, getDashboardData, renderPlainDashboard } from './dashboard-data.js';
 
-/**
- * Check worktree health by verifying paths exist on disk.
- */
-function checkWorktreeHealth(dispatches) {
-  return dispatches.map(d => ({
-    ...d,
-    healthy: d.worktreePath ? existsSync(d.worktreePath) : false,
-  }));
-}
-
-/**
- * Compute summary counts from dispatches.
- */
-export function computeSummary(dispatches) {
-  let active = 0;
-  let done = 0;
-  let blocked = 0;
-
-  for (const d of dispatches) {
-    if (d.status === 'done' || d.status === 'cleaned') {
-      done++;
-    } else if (!d.healthy) {
-      blocked++;
-    } else {
-      active++;
-    }
-  }
-
-  return { active, done, blocked };
-}
-
-/**
- * Load and process dashboard data.
- */
-export function getDashboardData({ project } = {}) {
-  const raw = getActiveDispatches();
-  let dispatches = checkWorktreeHealth(raw);
-
-  if (project) {
-    dispatches = dispatches.filter(d => d.repo && d.repo.includes(project));
-  }
-
-  const summary = computeSummary(dispatches);
-  return { dispatches, summary };
-}
+export { computeSummary, getDashboardData, renderPlainDashboard };
 
 /**
  * Summary line component.
@@ -66,50 +20,6 @@ function SummaryLine({ summary }) {
       </Text>
     </Box>
   );
-}
-
-/**
- * Render a plain text dashboard (no ANSI escape codes) for non-TTY environments.
- */
-export function renderPlainDashboard({ project } = {}) {
-  const { dispatches, summary } = getDashboardData({ project });
-  
-  const lines = [];
-  lines.push('Rally Dashboard');
-  lines.push('');
-  
-  // Table headers
-  const colWidths = { project: 20, issueRef: 12, branch: 28, status: 16, age: 6 };
-  const header = [
-    'Project'.padEnd(colWidths.project),
-    'Issue/PR'.padEnd(colWidths.issueRef),
-    'Branch'.padEnd(colWidths.branch),
-    'Status'.padEnd(colWidths.status),
-    'Age'.padEnd(colWidths.age),
-  ].join(' ');
-  lines.push(header);
-  
-  if (dispatches.length === 0) {
-    lines.push('No active dispatches');
-  } else {
-    for (const d of dispatches) {
-      const issueRef = d.type === 'pr' ? `PR #${d.number}` : `Issue #${d.number}`;
-      const age = formatAge(d.created ?? d.created_at);
-      const row = [
-        (d.repo ?? '').padEnd(colWidths.project),
-        issueRef.padEnd(colWidths.issueRef),
-        (d.branch ?? '').padEnd(colWidths.branch),
-        (d.status ?? '').padEnd(colWidths.status),
-        age.padEnd(colWidths.age),
-      ].join(' ');
-      lines.push(row);
-    }
-  }
-  
-  lines.push('');
-  lines.push(`${summary.active} active · ${summary.done} done · ${summary.blocked} blocked`);
-  
-  return lines.join('\n');
 }
 
 /**
