+++ lib/ui/index.js
@@ -1,4 +1,5 @@
 export { default as StatusMessage } from './components/StatusMessage.js';
 export { default as DispatchBox } from './components/DispatchBox.js';
-export { default as DispatchTable, STATUS_ICONS, formatAge } from './components/DispatchTable.js';
-export { default as Dashboard, getDashboardData, computeSummary, renderPlainDashboard } from './Dashboard.js';
+export { default as DispatchTable, STATUS_ICONS } from './components/DispatchTable.js';
+export { formatAge, getDashboardData, computeSummary, renderPlainDashboard } from './dashboard-data.js';
+export { default as Dashboard } from './Dashboard.js';
